+++ Robust.Shared/Toolshed/ToolshedManager.Permissions.cs
@@ -2,6 +2,12 @@
 
 public sealed partial class ToolshedManager
 {
+    /// <summary>
+    ///     The active permission controller, if any.
+    /// </summary>
+    /// <remarks>
+    ///     Invocation contexts can entirely ignore this, though it's bad form to do so if they have a session on hand.
+    /// </remarks>
     public IPermissionController? ActivePermissionController { get; private set; }
 
     public void SetPermissionController(IPermissionController controller)