+++ Robust.Shared/Toolshed/ToolshedManager.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿#pragma warning restore CS1591
+
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
@@ -8,6 +10,7 @@ using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
 using Robust.Shared.Log;
 using Robust.Shared.Network;
+using Robust.Shared.Players;
 using Robust.Shared.Reflection;
 using Robust.Shared.Timing;
 using Robust.Shared.Toolshed.Invocation;
@@ -17,6 +20,9 @@ using Robust.Shared.Utility;
 
 namespace Robust.Shared.Toolshed;
 
+/// <summary>
+///     The overarching controller for Toolshed, providing invocation, reflection, commands, parsing, and other tools used by the language.
+/// </summary>
 public sealed partial class ToolshedManager
 {
     [Dependency] private readonly IConsoleHost _conHost = default!;
@@ -30,6 +36,9 @@ public sealed partial class ToolshedManager
 
     private readonly Dictionary<string, ToolshedCommand> _commands = new();
 
+    /// <summary>
+    ///     If you're not an engine developer, you probably shouldn't call this.
+    /// </summary>
     public void Initialize()
     {
 #if !CLIENT_SCRIPTING
@@ -78,6 +87,10 @@ public sealed partial class ToolshedManager
         return result;
     }
 
+    /// <summary>
+    ///     Provides every registered command, including subcommands.
+    /// </summary>
+    /// <returns>Enumerable of every command.</returns>
     public IEnumerable<CommandSpec> AllCommands()
     {
         foreach (var (_, cmd) in _commands)
@@ -96,8 +109,20 @@ public sealed partial class ToolshedManager
         }
     }
 
+    /// <summary>
+    ///     Gets a command's object by name.
+    /// </summary>
+    /// <param name="commandName">Command to get.</param>
+    /// <returns>A command object.</returns>
+    /// <exception cref="IndexOutOfRangeException">Thrown when there is no command of the given name.</exception>
     public ToolshedCommand GetCommand(string commandName) => _commands[commandName];
 
+    /// <summary>
+    ///     Attempts to get a command's object by name.
+    /// </summary>
+    /// <param name="commandName">Command to get.</param>
+    /// <param name="command">The command obtained, if any.</param>
+    /// <returns>Success.</returns>
     public bool TryGetCommand(string commandName, [NotNullWhen(true)] out ToolshedCommand? command)
     {
         return _commands.TryGetValue(commandName, out command);
@@ -129,14 +154,48 @@ public sealed partial class ToolshedManager
                 }
             }
 
-            ctx.ClearErrors();
         }
 
         shell.WriteLine(FormattedMessage.FromMarkup(PrettyPrintType(result)));
     }
 
+    /// <summary>
+    ///     Invokes a command as the given user.
+    /// </summary>
+    /// <param name="session">User to run as.</param>
+    /// <param name="command">Command to invoke.</param>
+    /// <param name="input">An input value to use, if any.</param>
+    /// <param name="result">The resulting value, if any.</param>
+    /// <returns>Invocation success.</returns>
+    /// <remarks>
+    ///     This will use the same IInvocationContext as the one used by the user for debug console commands.
+    /// </remarks>
+    public bool InvokeCommand(ICommonSession session, string command, object? input, out object? result)
+    {
+        if (!_contexts.TryGetValue(session.UserId, out var ctx))
+        {
+            // Can't get a shell here.
+            result = null;
+            return false;
+        }
+
+        ctx.ClearErrors();
+
+        return InvokeCommand(ctx, command, input, out result);
+    }
+
+    /// <summary>
+    ///     Invokes a command with the given context.
+    /// </summary>
+    /// <param name="ctx">The context to run in.</param>
+    /// <param name="command">Command to invoke.</param>
+    /// <param name="input">An input value to use, if any.</param>
+    /// <param name="result">The resulting value, if any.</param>
+    /// <returns>Invocation success.</returns>
     public bool InvokeCommand(IInvocationContext ctx, string command, object? input, out object? result)
     {
+        ctx.ClearErrors();
+
         var parser = new ForwardParser(command, this);
         if (!CommandRun.TryParse(false, false, parser, input?.GetType(), null, false, out var expr, out _, out var err) || parser.Index < parser.MaxIndex)
         {
@@ -153,13 +212,22 @@ public sealed partial class ToolshedManager
     }
 }
 
+/// <summary>
+///     A command specification, containing both the command object and the selected subcommand if any.
+/// </summary>
+/// <param name="Cmd">Command object.</param>
+/// <param name="SubCommand">Subcommand, if any.</param>
 public readonly record struct CommandSpec(ToolshedCommand Cmd, string? SubCommand) : IAsType<ToolshedCommand>
 {
+    /// <inheritdoc/>
     public ToolshedCommand AsType()
     {
         return Cmd;
     }
 
+    /// <summary>
+    ///     Returns a completion option for this command, for use in autocomplete.
+    /// </summary>
     public CompletionOption AsCompletion()
     {
         return new CompletionOption(
@@ -168,10 +236,17 @@ public readonly record struct CommandSpec(ToolshedCommand Cmd, string? SubComman
             );
     }
 
+    /// <summary>
+    ///     Returns the full name of the command.
+    /// </summary>
     public string FullName() => $"{Cmd.Name}{(SubCommand is not null ? ":" + SubCommand : "")}";
 
+    /// <summary>
+    ///     Returns the localization string for the description of this command.
+    /// </summary>
     public string DescLocStr() => Cmd.UnlocalizedDescription(SubCommand);
 
+    /// <inheritdoc/>
     public override string ToString()
     {
         return Cmd.GetHelp(SubCommand);